Redone: 'DoubleInputForm.razor' is redone from scartch and now saves correctly.
Redone: UI fixes meant to get rid of reduntant code for input of 'ContentName', and 'TextInputsJson'. It is now a single input field that dynamically changes label and data meant to be saved from its input instead of having 3 different labels, input fields.

Added: extensive comments to explain the code